+++ tests/test_04_gitignore.py
@@ -7,11 +7,14 @@ import unittest
 from pathspec.gitignore import (
 	GitIgnoreSpec)
 
+from .util import (
+	debug_results,
+	get_includes)
+
 
 class GitIgnoreSpecTest(unittest.TestCase):
 	"""
-	The :class:`GitIgnoreSpecTest` class tests the :class:`.GitIgnoreSpec`
-	class.
+	The :class:`GitIgnoreSpecTest` class tests the :class:`.GitIgnoreSpec` class.
 	"""
 
 	def test_01_reversed_args(self):
@@ -19,13 +22,18 @@ class GitIgnoreSpecTest(unittest.TestCase):
 		Test reversed args for `.from_lines()`.
 		"""
 		spec = GitIgnoreSpec.from_lines('gitwildmatch', ['*.txt'])
-		results = set(spec.match_files([
+		files = {
 			'a.txt',
 			'b.bin',
-		]))
-		self.assertEqual(results, {
+		}
+
+		results = list(spec.check_files(files))
+		ignores = get_includes(results)
+		debug = debug_results(spec, results)
+
+		self.assertEqual(ignores, {
 			'a.txt',
-		})
+		}, debug)
 
 	def test_02_dir_exclusions(self):
 		"""
@@ -43,17 +51,21 @@ class GitIgnoreSpecTest(unittest.TestCase):
 			'test2/b.bin',
 			'test2/c/c.txt',
 		}
-		ignores = set(spec.match_files(files))
+
+		results = list(spec.check_files(files))
+		ignores = get_includes(results)
+		debug = debug_results(spec, results)
+
 		self.assertEqual(ignores, {
 			'test1/a.txt',
 			'test1/c/c.txt',
 			'test2/a.txt',
 			'test2/c/c.txt',
-		})
+		}, debug)
 		self.assertEqual(files - ignores, {
 			'test1/b.bin',
 			'test2/b.bin',
-		})
+		}, debug)
 
 	def test_02_file_exclusions(self):
 		"""
@@ -71,22 +83,26 @@ class GitIgnoreSpecTest(unittest.TestCase):
 			'Y/b.txt',
 			'Y/Z/c.txt',
 		}
-		ignores = set(spec.match_files(files))
+
+		results = list(spec.check_files(files))
+		ignores = get_includes(results)
+		debug = debug_results(spec, results)
+
 		self.assertEqual(ignores, {
 			'X/a.txt',
 			'X/Z/c.txt',
 			'Y/a.txt',
 			'Y/Z/c.txt',
-		})
+		}, debug)
 		self.assertEqual(files - ignores, {
 			'X/b.txt',
 			'Y/b.txt',
-		})
+		}, debug)
 
 	def test_02_issue_41_a(self):
 		"""
-		Test including a file and excluding a directory with the same name
-		pattern, scenario A.
+		Test including a file and excluding a directory with the same name pattern,
+		scenario A.
 		"""
 		# Confirmed results with git (v2.42.0).
 		spec = GitIgnoreSpec.from_lines([
@@ -94,31 +110,35 @@ class GitIgnoreSpecTest(unittest.TestCase):
 			'!*.yaml/',
 		])
 		files = {
-			'dir.yaml/file.sql',  # -
+			'dir.yaml/file.sql',   # -
 			'dir.yaml/file.yaml',  # 1:*.yaml
 			'dir.yaml/index.txt',  # -
-			'dir/file.sql',  # -
-			'dir/file.yaml',  # 1:*.yaml
-			'dir/index.txt',  # -
-			'file.yaml',  # 1:*.yaml
+			'dir/file.sql',        # -
+			'dir/file.yaml',       # 1:*.yaml
+			'dir/index.txt',       # -
+			'file.yaml',           # 1:*.yaml
 		}
-		ignores = set(spec.match_files(files))
+
+		results = list(spec.check_files(files))
+		ignores = get_includes(results)
+		debug = debug_results(spec, results)
+
 		self.assertEqual(ignores, {
 			'dir.yaml/file.yaml',
 			'dir/file.yaml',
 			'file.yaml',
-		})
+		}, debug)
 		self.assertEqual(files - ignores, {
 			'dir.yaml/file.sql',
 			'dir.yaml/index.txt',
 			'dir/file.sql',
 			'dir/index.txt',
-		})
+		}, debug)
 
 	def test_02_issue_41_b(self):
 		"""
-		Test including a file and excluding a directory with the same name
-		pattern, scenario B.
+		Test including a file and excluding a directory with the same name pattern,
+		scenario B.
 		"""
 		# Confirmed results with git (v2.42.0).
 		spec = GitIgnoreSpec.from_lines([
@@ -126,31 +146,35 @@ class GitIgnoreSpecTest(unittest.TestCase):
 			'*.yaml',
 		])
 		files = {
-			'dir.yaml/file.sql',  # 2:*.yaml
+			'dir.yaml/file.sql',   # 2:*.yaml
 			'dir.yaml/file.yaml',  # 2:*.yaml
 			'dir.yaml/index.txt',  # 2:*.yaml
-			'dir/file.sql',  # -
-			'dir/file.yaml',  # 2:*.yaml
-			'dir/index.txt',  # -
-			'file.yaml',  # 2:*.yaml
+			'dir/file.sql',        # -
+			'dir/file.yaml',       # 2:*.yaml
+			'dir/index.txt',       # -
+			'file.yaml',           # 2:*.yaml
 		}
-		ignores = set(spec.match_files(files))
+
+		results = list(spec.check_files(files))
+		ignores = get_includes(results)
+		debug = debug_results(spec, results)
+
 		self.assertEqual(ignores, {
 			'dir.yaml/file.sql',
 			'dir.yaml/file.yaml',
 			'dir.yaml/index.txt',
 			'dir/file.yaml',
 			'file.yaml',
-		})
+		}, debug)
 		self.assertEqual(files - ignores, {
 			'dir/file.sql',
 			'dir/index.txt',
-		})
+		}, debug)
 
 	def test_02_issue_41_c(self):
 		"""
-		Test including a file and excluding a directory with the same name
-		pattern, scenario C.
+		Test including a file and excluding a directory with the same name pattern,
+		scenario C.
 		"""
 		# Confirmed results with git (v2.42.0).
 		spec = GitIgnoreSpec.from_lines([
@@ -158,26 +182,30 @@ class GitIgnoreSpecTest(unittest.TestCase):
 			'!dir.yaml',
 		])
 		files = {
-			'dir.yaml/file.sql',  # -
+			'dir.yaml/file.sql',   # -
 			'dir.yaml/file.yaml',  # 1:*.yaml
 			'dir.yaml/index.txt',  # -
-			'dir/file.sql',  # -
-			'dir/file.yaml',  # 1:*.yaml
-			'dir/index.txt',  # -
-			'file.yaml',  # 1:*.yaml
+			'dir/file.sql',        # -
+			'dir/file.yaml',       # 1:*.yaml
+			'dir/index.txt',       # -
+			'file.yaml',           # 1:*.yaml
 		}
-		ignores = set(spec.match_files(files))
+
+		results = list(spec.check_files(files))
+		ignores = get_includes(results)
+		debug = debug_results(spec, results)
+
 		self.assertEqual(ignores, {
 			'dir.yaml/file.yaml',
 			'dir/file.yaml',
 			'file.yaml',
-		})
+		}, debug)
 		self.assertEqual(files - ignores, {
 			'dir.yaml/file.sql',
 			'dir.yaml/index.txt',
 			'dir/file.sql',
 			'dir/index.txt',
-		})
+		}, debug)
 
 	def test_03_subdir(self):
 		"""
@@ -195,23 +223,26 @@ class GitIgnoreSpecTest(unittest.TestCase):
 			'dirG/dirH/fileJ',
 			'dirG/fileO',
 		}
-		ignores = set(spec.match_files(files))
+
+		results = list(spec.check_files(files))
+		ignores = get_includes(results)
+		debug = debug_results(spec, results)
+
 		self.assertEqual(ignores, {
 			'dirG/dirH/fileI',
 			'dirG/dirH/fileJ',
 			'dirG/fileO',
-		})
+		}, debug)
 		self.assertEqual(files - ignores, {
 			'fileA',
 			'fileB',
 			'dirD/fileE',
 			'dirD/fileF',
-		})
+		}, debug)
 
 	def test_03_issue_19_a(self):
 		"""
-		Test matching files in a subdirectory of an included directory,
-		scenario A.
+		Test matching files in a subdirectory of an included directory, scenario A.
 		"""
 		spec = GitIgnoreSpec.from_lines([
 			"dirG/",
@@ -225,23 +256,26 @@ class GitIgnoreSpecTest(unittest.TestCase):
 			'dirG/dirH/fileJ',
 			'dirG/fileO',
 		}
-		ignores = set(spec.match_files(files))
+
+		results = list(spec.check_files(files))
+		ignores = get_includes(results)
+		debug = debug_results(spec, results)
+
 		self.assertEqual(ignores, {
 			'dirG/dirH/fileI',
 			'dirG/dirH/fileJ',
 			'dirG/fileO',
-		})
+		}, debug)
 		self.assertEqual(files - ignores, {
 			'fileA',
 			'fileB',
 			'dirD/fileE',
 			'dirD/fileF',
-		})
+		}, debug)
 
 	def test_03_issue_19_b(self):
 		"""
-		Test matching files in a subdirectory of an included directory,
-		scenario B.
+		Test matching files in a subdirectory of an included directory, scenario B.
 		"""
 		spec = GitIgnoreSpec.from_lines([
 			"dirG/*",
@@ -255,23 +289,26 @@ class GitIgnoreSpecTest(unittest.TestCase):
 			'dirG/dirH/fileJ',
 			'dirG/fileO',
 		}
-		ignores = set(spec.match_files(files))
+
+		results = list(spec.check_files(files))
+		ignores = get_includes(results)
+		debug = debug_results(spec, results)
+
 		self.assertEqual(ignores, {
 			'dirG/dirH/fileI',
 			'dirG/dirH/fileJ',
 			'dirG/fileO',
-		})
+		}, debug)
 		self.assertEqual(files - ignores, {
 			'fileA',
 			'fileB',
 			'dirD/fileE',
 			'dirD/fileF',
-		})
+		}, debug)
 
 	def test_03_issue_19_c(self):
 		"""
-		Test matching files in a subdirectory of an included directory,
-		scenario C.
+		Test matching files in a subdirectory of an included directory, scenario C.
 		"""
 		spec = GitIgnoreSpec.from_lines([
 			"dirG/**",
@@ -285,18 +322,22 @@ class GitIgnoreSpecTest(unittest.TestCase):
 			'dirG/dirH/fileJ',
 			'dirG/fileO',
 		}
-		ignores = set(spec.match_files(files))
+
+		results = list(spec.check_files(files))
+		ignores = get_includes(results)
+		debug = debug_results(spec, results)
+
 		self.assertEqual(ignores, {
 			'dirG/dirH/fileI',
 			'dirG/dirH/fileJ',
 			'dirG/fileO',
-		})
+		}, debug)
 		self.assertEqual(files - ignores, {
 			'fileA',
 			'fileB',
 			'dirD/fileE',
 			'dirD/fileF',
-		})
+		}, debug)
 
 	def test_04_issue_62(self):
 		"""
@@ -306,14 +347,19 @@ class GitIgnoreSpecTest(unittest.TestCase):
 			'*',
 			'!product_dir/',
 		])
-		results = set(spec.match_files([
+		files = {
 			'anydir/file.txt',
 			'product_dir/file.txt',
-		]))
-		self.assertEqual(results, {
+		}
+
+		results = list(spec.check_files(files))
+		ignores = get_includes(results)
+		debug = debug_results(spec, results)
+
+		self.assertEqual(ignores, {
 			'anydir/file.txt',
 			'product_dir/file.txt',
-		})
+		}, debug)
 
 	def test_05_issue_39(self):
 		"""
@@ -331,16 +377,20 @@ class GitIgnoreSpecTest(unittest.TestCase):
 			'important/e.txt',
 			'trace.c',
 		}
-		ignores = set(spec.match_files(files))
+
+		results = list(spec.check_files(files))
+		ignores = get_includes(results)
+		debug = debug_results(spec, results)
+
 		self.assertEqual(ignores, {
 			'a.log',
 			'trace.c',
-		})
+		}, debug)
 		self.assertEqual(files - ignores, {
 			'b.txt',
 			'important/d.log',
 			'important/e.txt',
-		})
+		}, debug)
 
 	def test_06_issue_64(self):
 		"""
@@ -359,8 +409,12 @@ class GitIgnoreSpecTest(unittest.TestCase):
 			'A/B/C/x',
 			'A/B/C/y.py',
 		}
-		ignores = set(spec.match_files(files))
-		self.assertEqual(ignores, files)
+
+		results = list(spec.check_files(files))
+		ignores = get_includes(results)
+		debug = debug_results(spec, results)
+
+		self.assertEqual(ignores, files, debug)
 
 	def test_07_issue_74(self):
 		"""
@@ -380,21 +434,25 @@ class GitIgnoreSpecTest(unittest.TestCase):
 			'test2/b.bin',
 			'test2/c/c.txt',
 		}
-		ignores = set(spec.match_files(files))
+
+		results = list(spec.check_files(files))
+		ignores = get_includes(results)
+		debug = debug_results(spec, results)
+
 		self.assertEqual(ignores, {
 			'test1/b.bin',
 			'test1/a.txt',
 			'test1/c/c.txt',
 			'test2/b.bin',
-		})
+		}, debug)
 		self.assertEqual(files - ignores, {
 			'test2/a.txt',
 			'test2/c/c.txt',
-		})
+		}, debug)
 
 	def test_08_issue_81_a(self):
 		"""
-		Test issue 81, scenario A.
+		Test issue 81 whitelist, scenario A.
 		"""
 		# Confirmed results with git (v2.42.0).
 		spec = GitIgnoreSpec.from_lines([
@@ -403,20 +461,24 @@ class GitIgnoreSpecTest(unittest.TestCase):
 			"!libfoo/**",
 		])
 		files = {
-			"ignore.txt",  # 1:*
+			"ignore.txt",          # 1:*
 			"libfoo/__init__.py",  # 3:!libfoo/**
 		}
-		ignores = set(spec.match_files(files))
+
+		results = list(spec.check_files(files))
+		ignores = get_includes(results)
+		debug = debug_results(spec, results)
+
 		self.assertEqual(ignores, {
 			"ignore.txt",
-		})
+		}, debug)
 		self.assertEqual(files - ignores, {
 			"libfoo/__init__.py",
-		})
+		}, debug)
 
 	def test_08_issue_81_b(self):
 		"""
-		Test issue 81, scenario B.
+		Test issue 81 whitelist, scenario B.
 		"""
 		# Confirmed results with git (v2.42.0).
 		spec = GitIgnoreSpec.from_lines([
@@ -425,20 +487,24 @@ class GitIgnoreSpecTest(unittest.TestCase):
 			"!libfoo/*",
 		])
 		files = {
-			"ignore.txt",  # 1:*
+			"ignore.txt",          # 1:*
 			"libfoo/__init__.py",  # 3:!libfoo/*
 		}
-		ignores = set(spec.match_files(files))
+
+		results = list(spec.check_files(files))
+		ignores = get_includes(results)
+		debug = debug_results(spec, results)
+
 		self.assertEqual(ignores, {
 			"ignore.txt",
-		})
+		}, debug)
 		self.assertEqual(files - ignores, {
 			"libfoo/__init__.py",
-		})
+		}, debug)
 
 	def test_08_issue_81_c(self):
 		"""
-		Test issue 81, scenario C.
+		Test issue 81 whitelist, scenario C.
 		"""
 		# Confirmed results with git (v2.42.0).
 		spec = GitIgnoreSpec.from_lines([
@@ -447,12 +513,14 @@ class GitIgnoreSpecTest(unittest.TestCase):
 			"!libfoo/",
 		])
 		files = {
-			"ignore.txt",  # 1:*
+			"ignore.txt",          # 1:*
 			"libfoo/__init__.py",  # 1:*
 		}
-		ignores = set(spec.match_files(files))
+		results = list(spec.check_files(files))
+		ignores = get_includes(results)
+		debug = debug_results(spec, results)
 		self.assertEqual(ignores, {
 			"ignore.txt",
 			"libfoo/__init__.py",
-		})
+		}, debug)
 		self.assertEqual(files - ignores, set())